+++ apps/api/src/app/portfolio/calculator/mwr/portfolio-calculator.ts
@@ -1,10 +1,6 @@
 import { PortfolioCalculator } from '@ghostfolio/api/app/portfolio/calculator/portfolio-calculator';
-import { PortfolioSnapshot } from '@ghostfolio/api/app/portfolio/interfaces/portfolio-snapshot.interface';
-import {
-  SymbolMetrics,
-  TimelinePosition,
-  UniqueAsset
-} from '@ghostfolio/common/interfaces';
+import { SymbolMetrics, UniqueAsset } from '@ghostfolio/common/interfaces';
+import { PortfolioSnapshot, TimelinePosition } from '@ghostfolio/common/models';
 
 export class MWRPortfolioCalculator extends PortfolioCalculator {
   protected calculateOverallPerformance(