+++ apps/api/src/app/portfolio/calculator/portfolio-calculator.factory.ts
@@ -1,8 +1,10 @@
 import { Activity } from '@ghostfolio/api/app/order/interfaces/activities.interface';
 import { CurrentRateService } from '@ghostfolio/api/app/portfolio/current-rate.service';
+import { RedisCacheService } from '@ghostfolio/api/app/redis-cache/redis-cache.service';
+import { ConfigurationService } from '@ghostfolio/api/services/configuration/configuration.service';
 import { ExchangeRateDataService } from '@ghostfolio/api/services/exchange-rate-data/exchange-rate-data.service';
 import { HistoricalDataItem } from '@ghostfolio/common/interfaces';
-import { DateRange } from '@ghostfolio/common/types';
+import { DateRange, UserWithSettings } from '@ghostfolio/common/types';
 
 import { Injectable } from '@nestjs/common';
 
@@ -18,8 +20,10 @@ export enum PerformanceCalculationType {
 @Injectable()
 export class PortfolioCalculatorFactory {
   public constructor(
+    private readonly configurationService: ConfigurationService,
     private readonly currentRateService: CurrentRateService,
-    private readonly exchangeRateDataService: ExchangeRateDataService
+    private readonly exchangeRateDataService: ExchangeRateDataService,
+    private readonly redisCacheService: RedisCacheService
   ) {}
 
   public createCalculator({
@@ -27,13 +31,17 @@ export class PortfolioCalculatorFactory {
     activities,
     calculationType,
     currency,
-    dateRange = 'max'
+    dateRange = 'max',
+    isExperimentalFeatures = false,
+    userId
   }: {
     accountBalanceItems?: HistoricalDataItem[];
     activities: Activity[];
     calculationType: PerformanceCalculationType;
     currency: string;
     dateRange?: DateRange;
+    isExperimentalFeatures?: boolean;
+    userId: string;
   }): PortfolioCalculator {
     switch (calculationType) {
       case PerformanceCalculationType.MWR:
@@ -42,8 +50,12 @@ export class PortfolioCalculatorFactory {
           activities,
           currency,
           dateRange,
+          isExperimentalFeatures,
+          userId,
+          configurationService: this.configurationService,
           currentRateService: this.currentRateService,
-          exchangeRateDataService: this.exchangeRateDataService
+          exchangeRateDataService: this.exchangeRateDataService,
+          redisCacheService: this.redisCacheService
         });
       case PerformanceCalculationType.TWR:
         return new TWRPortfolioCalculator({
@@ -52,7 +64,11 @@ export class PortfolioCalculatorFactory {
           currency,
           currentRateService: this.currentRateService,
           dateRange,
-          exchangeRateDataService: this.exchangeRateDataService
+          isExperimentalFeatures,
+          userId,
+          configurationService: this.configurationService,
+          exchangeRateDataService: this.exchangeRateDataService,
+          redisCacheService: this.redisCacheService
         });
       default:
         throw new Error('Invalid calculation type');