+++ apps/api/src/app/portfolio/calculator/twr/portfolio-calculator-googl-buy.spec.ts
@@ -1,7 +1,8 @@
 import { Activity } from '@ghostfolio/api/app/order/interfaces/activities.interface';
 import {
   activityDummyData,
-  symbolProfileDummyData
+  symbolProfileDummyData,
+  userDummyData
 } from '@ghostfolio/api/app/portfolio/calculator/portfolio-calculator-test-utils';
 import {
   PortfolioCalculatorFactory,
@@ -9,6 +10,9 @@ import {
 } from '@ghostfolio/api/app/portfolio/calculator/portfolio-calculator.factory';
 import { CurrentRateService } from '@ghostfolio/api/app/portfolio/current-rate.service';
 import { CurrentRateServiceMock } from '@ghostfolio/api/app/portfolio/current-rate.service.mock';
+import { RedisCacheService } from '@ghostfolio/api/app/redis-cache/redis-cache.service';
+import { RedisCacheServiceMock } from '@ghostfolio/api/app/redis-cache/redis-cache.service.mock';
+import { ConfigurationService } from '@ghostfolio/api/services/configuration/configuration.service';
 import { ExchangeRateDataService } from '@ghostfolio/api/services/exchange-rate-data/exchange-rate-data.service';
 import { ExchangeRateDataServiceMock } from '@ghostfolio/api/services/exchange-rate-data/exchange-rate-data.service.mock';
 import { parseDate } from '@ghostfolio/common/helper';
@@ -20,6 +24,15 @@ jest.mock('@ghostfolio/api/app/portfolio/current-rate.service', () => {
     // eslint-disable-next-line @typescript-eslint/naming-convention
     CurrentRateService: jest.fn().mockImplementation(() => {
       return CurrentRateServiceMock;
+    })
+  };
+});
+
+jest.mock('@ghostfolio/api/app/redis-cache/redis-cache.service', () => {
+  return {
+    // eslint-disable-next-line @typescript-eslint/naming-convention
+    RedisCacheService: jest.fn().mockImplementation(() => {
+      return RedisCacheServiceMock;
     })
   };
 });
@@ -37,11 +50,15 @@ jest.mock(
 );
 
 describe('PortfolioCalculator', () => {
+  let configurationService: ConfigurationService;
   let currentRateService: CurrentRateService;
   let exchangeRateDataService: ExchangeRateDataService;
   let factory: PortfolioCalculatorFactory;
+  let redisCacheService: RedisCacheService;
 
   beforeEach(() => {
+    configurationService = new ConfigurationService();
+
     currentRateService = new CurrentRateService(null, null, null, null);
 
     exchangeRateDataService = new ExchangeRateDataService(
@@ -51,9 +68,13 @@ describe('PortfolioCalculator', () => {
       null
     );
 
+    redisCacheService = new RedisCacheService(null, null);
+
     factory = new PortfolioCalculatorFactory(
+      configurationService,
       currentRateService,
-      exchangeRateDataService
+      exchangeRateDataService,
+      redisCacheService
     );
   });
 
@@ -84,7 +105,8 @@ describe('PortfolioCalculator', () => {
       const portfolioCalculator = factory.createCalculator({
         activities,
         calculationType: PerformanceCalculationType.TWR,
-        currency: 'CHF'
+        currency: 'CHF',
+        userId: userDummyData.id
       });
 
       const chartData = await portfolioCalculator.getChartData({