+++ apps/api/src/app/portfolio/portfolio.service.ts
@@ -29,6 +29,7 @@ import {
   EnhancedSymbolProfile,
   Filter,
   HistoricalDataItem,
+  InvestmentItem,
   PortfolioDetails,
   PortfolioInvestments,
   PortfolioPerformanceResponse,
@@ -36,10 +37,9 @@ import {
   PortfolioReport,
   PortfolioSummary,
   Position,
-  TimelinePosition,
   UserSettings
 } from '@ghostfolio/common/interfaces';
-import { InvestmentItem } from '@ghostfolio/common/interfaces/investment-item.interface';
+import { TimelinePosition } from '@ghostfolio/common/models';
 import type {
   AccountWithValue,
   DateRange,
@@ -277,8 +277,11 @@ export class PortfolioService {
 
     const portfolioCalculator = this.calculatorFactory.createCalculator({
       activities,
+      userId,
       calculationType: PerformanceCalculationType.TWR,
-      currency: this.request.user.Settings.settings.baseCurrency
+      currency: this.request.user.Settings.settings.baseCurrency,
+      isExperimentalFeatures:
+        this.request.user.Settings.settings.isExperimentalFeatures
     });
 
     const items = await portfolioCalculator.getChart({
@@ -352,8 +355,11 @@ export class PortfolioService {
     const portfolioCalculator = this.calculatorFactory.createCalculator({
       activities,
       dateRange,
+      userId,
       calculationType: PerformanceCalculationType.TWR,
-      currency: userCurrency
+      currency: userCurrency,
+      isExperimentalFeatures:
+        this.request.user.Settings.settings.isExperimentalFeatures
     });
 
     const { currentValueInBaseCurrency, hasErrors, positions } =
@@ -648,11 +654,14 @@ export class PortfolioService {
     ]);
 
     const portfolioCalculator = this.calculatorFactory.createCalculator({
+      userId,
       activities: orders.filter((order) => {
         return ['BUY', 'DIVIDEND', 'ITEM', 'SELL'].includes(order.type);
       }),
       calculationType: PerformanceCalculationType.TWR,
-      currency: userCurrency
+      currency: userCurrency,
+      isExperimentalFeatures:
+        this.request.user.Settings.settings.isExperimentalFeatures
     });
 
     const portfolioStart = portfolioCalculator.getStartDate();
@@ -919,8 +928,11 @@ export class PortfolioService {
     const portfolioCalculator = this.calculatorFactory.createCalculator({
       activities,
       dateRange,
+      userId,
       calculationType: PerformanceCalculationType.TWR,
-      currency: this.request.user.Settings.settings.baseCurrency
+      currency: this.request.user.Settings.settings.baseCurrency,
+      isExperimentalFeatures:
+        this.request.user.Settings.settings.isExperimentalFeatures
     });
 
     let { hasErrors, positions } = await portfolioCalculator.getSnapshot();
@@ -1108,8 +1120,11 @@ export class PortfolioService {
       accountBalanceItems,
       activities,
       dateRange,
+      userId,
       calculationType: PerformanceCalculationType.TWR,
-      currency: userCurrency
+      currency: userCurrency,
+      isExperimentalFeatures:
+        this.request.user.Settings.settings.isExperimentalFeatures
     });
 
     const {
@@ -1202,8 +1217,11 @@ export class PortfolioService {
 
     const portfolioCalculator = this.calculatorFactory.createCalculator({
       activities,
+      userId,
       calculationType: PerformanceCalculationType.TWR,
-      currency: this.request.user.Settings.settings.baseCurrency
+      currency: this.request.user.Settings.settings.baseCurrency,
+      isExperimentalFeatures:
+        this.request.user.Settings.settings.isExperimentalFeatures
     });
 
     let { totalFeesWithCurrencyEffect, positions, totalInvestment } =