+++ libs/common/src/lib/class-transformer.ts
@@ -1,0 +1,9 @@
+import { Big } from 'big.js';
+
+export function transformToBig({ value }: { value: string }): Big {
+  if (value === null) {
+    return null;
+  }
+
+  return new Big(value);
+}